SimuLine/Preprocessing/UnbiasedEmbedding/requirements.txt
numpy==1.18.1
pandas==0.25.1
scikit-learn==0.23.1
tensorflow==1.15.2
pyyaml==5.1.2